Change project structure
https://monorepo.guide/getting-started suggests that apps directory
should contain "user-facing apps and websites" and packages directory
should contain "packages designed to be consumed by other packages
(published OR internal)". I'm not sure if this is a common convention,
but it lookes simple and clear enough for me.

Rename our nextjs app to @thunderstore/nextjs and components package
to @thunderstore/components. Fix imports to reflect the changes.

Update the references in GitHub Actions files to reflect the new file
paths.