Fix Markdown component
Introducing Preconstruct caused some weird issues. Whenever
"react-markdown" or "remark-gfm" is imported in a normal way, the app
crashes with "Error: Element type is invalid: expected a string (for
built-in components) or a class/function (for composite components) but
got: undefined". Furthermore, the error only occurs after a full page
refresh - regular imports don't cause problems when changes are updated
by live-reload.

The same error occurs with the package versions used before this PR as
well as with the latest ones.