Move dependencies between different package.json files
As far as I can tell, the recommendation is that each app and package
in the monorepo defines all the dependencies it uses. Yarn workspaces
will then handle hoisting and deduplicating the dependencies. Root
level package.json should contain the tooling dependencies used by the
project as a whole. Also I believe this only matters in cases where
packages are going to be published separately. I didn't manage to find
any solid information on this.