Fix get-node-version GitHub Action
Update the action to read all versions from dependencies, since the
root package.json no longer contains a devDependencies section. Update
the packages in the action to use the same versions that are used in
the actual repo, and then update the Docker image to use the same Node
version.