Initialize Storybook
Storybook is an open source tool for building UI components and pages
in isolation. In this project it hopefully helps us to create reusable
UI components.

Storybook was initialized using the CLI tool and react-webpack default
settings. Some dependencies were manually added to package.json to get
rid of warnings.