Remove line breaks from markdown
Markdown uses double trailing spaces as a line break. Pre-commit will
want to remove these. Remove the lines manually. This will make the
elements  be rendered next  to each other in the Storybook, but
figuring out how to keep the trailing whitespace in this one file
doesn't seem worith it.