@thunderstore/components: add TopBar
TopBar is a navigation element shown on top of every page.

The designs in Figma had only a single login link, but since we'll
support multiple OAuth providers, a drowdown menu with links was used
instead.

Since the authentication isn't implemented yet, the login links
redirect the user to frontpage. Also for the same reason the TopBar
doesn't currently adjust it contents for authenticated users.

Refs TS-189