@thunderstore/nextjs: change how page backgrounds are rendered
The app has two types of pages. First ones (e.g. the landing page) have
a taller image that is shown on top of the gradient, but the background
is shown only for the top of the page, where as the bottom of the page
has a flat single-color background. Second ones (e.g a community page)
have a short, banner-like image and display a full-page gradient on top
of the image.

The previous implementation was for the first type, and worked poorly
for the second type.

* Background and HalfPageBackground components can be used to add
  background image and radial gradient to any page.
* Content's max-width and centering styles are no longer defined as a
  part of Chakra UI's theme. Instead all centered content should be
  wrapped in ContentWrapper component. From separation-of-concerns
  point of view this is an improvement, since these stylistic choices
  belong more to the Next.js app than the component library.
* TopBar is no longer included in _app.tsx, meaning that it's not
  automatically shown on all pages anymore. Instead it's part of the
  ContentWrapper component.
* The CSS for generating the radial gradient was changed since what
  worked nicely on Figma, didn't work that nicely on actual pages. The
  size of the gradient is now defined in pixels instead of percents so
  it looks nice on different page heights/widths. Also the vertical
  positioning is now fixed in pixels so the center is always near the
  top of the page. Horizontal positioning is still in percents, since
  that works better on varying device widths.

Refs TS-189