@thunderstore/components: add TagBox
This is a simple component derived from Chakra UI's Box component, with
some additional styling. Custom component is used instead of creating a
variant of the Box component to avoid conflict when size props are
used.

EDIT: after making this commit I learned that Chakra also provides a
Tag components. It might make sense check out if this one should be
replaced with variant styles for that component.

Refs TS-201