@thunderstore/nextjs: add useQueryToState hook
This custom hook basically wraps to other hooks: first it uses
useRouter from Next.js to check if given key exists in route's GET
parameters. That value is then stored using React's useState.

The hook also casts the strings received from useRouter to boolean,
integer, or string. Additionally, arrays of integers and strings are
supported, in case the same parameter exists multiple times in the URL.
The typing is defined by passing one of provided casting functions when
the hook is initiated.