@thunderstore/components: upgrade @chakra-ui/react to 1.7.4
This adds support for columnGap/rowGap properties and brings the
components library to use the same version as the Next.js app.

Note 1: for me this broke down the columnGap support on the Next.js
app. Pruning node_modules solved the issue.

Note 2: a.k.a. where things get frustrating. See the PR for more
information.