Update @chakra-ui/react to 0.0.0-pr-2022062070
The library was updated to 1.7.4, but unfortunately this broke typings
of defaultProps & size and variant props related theming component
styles.

Version 1.7.5 (which isn't mentioned in the release notes, but is
available in NPM) doesn't fix the issue.

The version used in this commit is based on this PR:
https://github.com/chakra-ui/chakra-ui/pull/5342 . I assume we can
return to regular releases once 1.7.6 is released. While using
unsupported release is far from ideal, I want to move on with these
changes, and since the project isn't released yet, we can live on the
bleeding edge for a while.