Add URL query parameter support to LinkingProvider
Due to our LinkingProvider implementation which separates the abstract
links from their actual URLs, we can't easily manually append the
parameters at the end of the URL, as could be done with regular anchor
elements. Therefore the link components must accept the query
parameters as a prop. It's up to the app specific implementation how
the queryParams prop is actually utilized.

Additionally add "team" to thunderstoreLinkProps so it's not passed to
the actual anchor component as a prop. This was missing due to mistake.