@thunderstore/components: add PackageActions
PackageActions shows options for installing and downloading the package
on package's detail view. Also stats about downloads, likes and latest
update are shown.

At the bottom of the component there's options for copying the id
string of the package to clipboard (not implemented) and moving to
package's dependants view. However, these are WIP since we haven't
decided yat how the UI should actually look and function.

Refs TS-204