@thunderstore/components: add styles to content rendered by Markdown
The styles are a combination of what can be seen in Figma designs and
what's in use on the old website. While the styles probably aren't
comprehensive, I tried to check some mod pages to see how markdown is
currently used in the READMEs.

The styles are used via Chakra UI's global styles feature, meaning that
they'll apply to all divs with class name "markdown", whether the div
is rendered by the Markdown component or not.