@thunderstore/nextjs: fix misunderstanding regarding mods/modpacks
Initial implementation assumed that there's two types for packages:
mods and modpacks. This is not the case: modpacks are filtered by
using backend's PackageListingSection funtionality, which is a shortcut
for filtering packages by a set of required/exluded categories.

This implementation differs from the old UI in that neither mods nor
modpacks are selected by default, unless the GET parameters define
otherwise.