Add @thunderstore/dapper package
Dapper is a data provider solution which sits between a client and the
backend. It provides the client with strongly typed functions which
return the data required to e.g. render a specific page. The client
doesn't need to be conserned how and from where the data is acquited
from. Ideally Dapper also defines the interface for this data exchange:
new interchangeable implementations can be written for different
clients and data sources.

Dapper provides a Context component and a hook for easy access in React
components. Since getServerSideProps method of Next.js doesn't support
hooks, Dapper can also be initiated as a regular class.

As a proof-of-concept, this commit includes the method for fetching the
data required on community's package listing view.

Refs TS-229