@thunderstore/storybook: fix wandering human readable date
Story's content used a fixed date, which caused the human readable
output to wander from "2 months ago" to "3 months ago" and so on. This
triggered unnecessary and misleading need for a review in Chromatic.

The story now uses dynamic date, which causes the human readable
format to remain as "last week".