Move and refactor TopBar component
In hindsight, completely defining the TopBar and its contents under the
components library is a bad idea, since it makes it difficult to
implement the app specific dynamic functionality like logging in and
adjusting content based on users authentication status.

@thunderstore/components now provides a skeleton TopBar component,
which the apps can augment via children prop, as well as some
subcomponents which hopefully makes it easier for the app to define
those children.