@thunderstore/nextjs: validate session id read from localStorage
When user authenticates, the session id is stored in SessionProvider's
internal state for quick access and localStorage for retention. When
the session id is read from the localStorage, API backend is pinged to
check if the session is still valid.

Consumers of the SessionProvider context should check the provided
isSessionIdValidated flag before attempting to use the sessionId e.g.
in API calls.

Refs TS-230, TS-358