@thunderstore/sertra: add wrappers for server submit form's elements
The original goal was to fix the height of "Select Mods" button, which
got stretched due to parent's flex rules. Gkruck suggested adding
wrapper divs that control the layout and lets child elements to control
their own dimensions. It was also pointed out that this makes things
easier if we add more content to each row. That's a lot of divs to
adjust the height of one button though...