@thunderstore/sertra: store mod selection on page instead of modal
Move the list of selected mods from the state of mod selection modal to
the state of the page component. This is required for Done/Cancel
buttons to work properly (mod list is not updated until user clicks
done) as well as displaying the selected mods on the form without
having to parse the package name from the canonical name.

Since the selected mods are stored on the page, it also makes sense to
do the conversion from Packages to ModPackages in the page component.
Modal can interact with ModPackages without having to know anything
about the Packages.

Refs TS-1060