@thunderstore/sertra: sort mods by name on selection modal
This prevents the mod from changing place on the list when it's removed
and readded, and when the selected version number is changed.

It might also make it easier to find a specific mod from a long list,
although I'm not sure if the ordering returned from API was meaningful,
e.g. pinned mods first?

Refs TS-1060