Add stylelint
Add linting for CSS files to keep the codebase consistent and to avoid
pickering in code reviews.

* Add stylelint dependencies
* Configure pre-commit to run stylelint hooks
* Tell prettier to ignore CSS files. Otherwise prettier would do
  changes that would immediately be undone by stylelint, which would
  lead to pre-commit rejecting the commit. Using a premade stylelint
  config to disable rules that may conflict with prettier seemed to
  turn off too many rules, so this approach was chosen instead
* Adjust a pair of premade stylelint configs with some house rules.
  Mostly these are based on Antti's personal preferences. Class
  selector format was discussed for quite some time. Others preferred
  camelCase with short descriptive names, others preferred BEM
  notation. In the end it was decided to just blacklist hyphens,
  since using such selectors is awkward in JS side of CSS modules.

Note that currently @thunderstore/components isn't linted, since it
defines styles inside TS files and is otherwise heavily coupled with
Chakra UI. It was decided that we want to use CSS modules instead of
Chakra in the future, so style declarations in TS files will be phased
out eventually.

Refs TS-1219