CS: Implement SSR compatible dapper context
Implement a workaround for the lack of React Context support on the
NextJS serverside rendering process. Context isn't supported as it's
often used for state management, but in our case we just need a way to
provide the dapper implementation into some global scope which can later
be consumed by the components without direct references to the
implementation.

The implemented workaround relies on globalThis if it's available
(server, newer browsers) or window if it isn't (older browsers).
globalThis might persist throughout multiple requests served by the
NextJS server, meaning instance-wide serverside caching should be
possible to implement in dapper.