@thunderstore/cyberstorm: improve PackageListing category updating
- Check that `filters` and `datas` are defined before proceeding
  further, since anything can't be updated if they're not present,
  which should improve performance as well as clean up the code from
  non-null assertions
- Make a copy of `filters.availableCategories` instead of editing it
  in-place, as is recommended by React on things that are stored with
  useState hook
- use forEach instead of map when were not interested in what's
  returned
- Simplify how the category object is updated
- Replace `if (myObject)` check (which always evaluates to true) with a
  deep compare between the objects to see if the state needs to be
  updated

The changes also fix the issue where category filter list is empty
until user interacts with the search input.

This probably won't be the last we'll hear from this component, since I
have doubts that the implementation as a whole will work as intended
when there's multiple pages of real results instead of just dummy data.
Also the implementation currently uses React hooks inside a class,
which is a faux pass and will be flagged by upcoming linting changes.

Refs TS-1781