Fix illegal imports from @thunderstore/dapper
As mentioned in the previous commit, only stuff explicitly exported
Dapper package can now be imported elsewhere, although unfortunately at
least VSCode doesn't seem to know it so the errors are detected only by
CI pipeline when the packages are built.

Fix some of Dapper's internal imports as well to use relative imports
instead of using the package's name. This is based only on consistency
with the rest of the code base.