@thunderstore/cyberstorm: add suspense to CommunityListLayout
To get the suspense to work properly, the Suspense element needs to be
outside the asynchronous component, so CommunityList was extracted into
a separate component than handles the async Dapper operation.

Our current versions of dependencies, including TypeScript, doesn't
understand asynchronous components, and updating them is outside the
scope of this task, so the error is only silenced at this point.

Refs TS-1815