@thunderstore/thunderstore-api: add basic error handling to apiFetch
This is more for debugging purposes since we don't currently show the
errors to users. Later on Sentry should be plugged into the try-catch
blocks to provide us with more information about what's going on.

The old corresponding approach used custom DapperError for this, but I
decided to not use it this time to keep @thunderstore/thunderstore-api
free of references to @thunderstore/dapper. Neither of those packages
know of each others existence, it's @thunderstore/dapper-ts's job to
bring these two together.

Refs TS-1815