@thunderstore/dapper-ts: implement getCommunities
Pass the arguments to corresponding thunderstore-api method and use Zod
to validate the received data.

It was earlier suspected that Zod caused the dev server to run out of
memory, but based on some testing this no longer seems to be the case.
This might be due to upgraded Zod version, or maybe the problem doesn't
show itself with just one validator in place. Time will tell.

Refs TS-1815