Add @thunderstore/use-promise
react-promise-suspense used with Dapper methods has a critical flaw:
only the arguments of the function to be called are used as cache keys.
Thus using it to call two completely different functions with same
arguments (e.g. no arguments) causes one of them to get completely
wrong results from cache.

This flaw caused the builds of cyberstorm-nextjs to fail every now and
then when different Dapper methods got totally wrong results. This is
probably also the reason why dapper-fake sometimes returned odd results
for the Storybook.

As the package isn't actively maintained anymore, the external
dependency will be dropped in favour of a local implementation that
adds a quick'n'dirty fix for the issue.

Adding unit tests was considered, but it wasn't immediately obvious how
to use Jest to test a function that throws Promises. Tests can be added
later if deemed necessary, but for now it's more important to get the
build process that is a part of our CI pipeline to work reliably.

Refs no ticket