@thunderstore/cyberstorm: add enum support to Select
Select supported only values typed as string. The tasks request support
for string literals, and they are now supported, but it will require
type casting the string values of options as string literals, which
isn't very nice. However, the implementation also supports enums, which
leverage type inference and are much nicer to use.

Refs TS-1821