@thunderstore/cyberstorm: rename TeamLayout to TeamProfileLayout
Try to avoid confusion by renaming the layout to "team profile", akin
to UserProfileLayout that serves a similar function of acting as a
public view containing information about the team/user. Furthermore
this attemps to more clearly distinguish team profile from other
team-related views, which are used for managing the teams of the
authenticated user.

Refs TS-1855