@thunderstore/cyberstorm: drop Dapper call from TeamProfileLayout
The current content on the page only requires the team's name, and it
can be read from the URL. Trying to access non-existent team's profile
will (eventually) fail when fetching the team's packages fails, leading
the user to error page, so a separate checkup for that isn't needed
either.

Using team's id rather than the name to call PackageSearchLayout might
be a valid need for the id, but I'm not sure which way I'll end up
implementing the raleted Dapper method, so I don't consider that issue
yet.

Once we have more content to show on the profile page, a Dapper
request can be reinstated.

Refs TS-1856