@thunderstore/cyberstorm: avoid using attributes to be removed
Package and PackagePreview interfaces provided by Dapper currently
contains some duplicate fields that will be cleaned out. Author is not
needed, since packages are not authored by a user, but by a team. Team
doesn't have a separate human-readable name, so the id field will be
used. Both of these are and will be available by the namespace field,
which is used to name things consistently with the backend.

Refs TS-1860