@thunderstore/dapper: strongly type the scope of GetPackageListings
THe scope refers to the factors limiting the base result set when no
filters etc. are in play. E.g. the listing can be scoped to packages
listed in a specific community.

In attempt to be reusable, the earlier interface included a bunch of id
strings, e.g. community identifier and namespace name. All of these
were optional and there was no plan for which part of the code base
should be responsible to parsing the combination of these values to the
actual scope.

My initial idea was to do that parsing on the PackageList component,
but in the end I chose to do this at the top the chain in Layout
components. The rationale being that doing the parsing in middle of the
chain in unnecessary and a possible source of errors. It's better to
bypass it by requiring the Layouts to be explicit about the listing
from the get go.

It might seem that this couples Cyberstorm components and Dapper more
tightly together, but I'd argue the coupling was there already with the
old implementation. The changes in this commit only make that coupling
explicit.

Refs TS-1875