Fix package .gitignore plop template
The .gitignore used in a package plop template would not get
appropriately copied over to the newly created package. This seems to be
caused by plop expecting a filename to exist, which is not true for
dotfiles.

As a workaround, we set the filename to a variable that does not exist,
effectively leading to the end result we want.