@thunderstore/thunderstore-api: add two package listing fetchers
One will fetch package listings scoped on a specific community, while
the other scopes for both community and namespace.

Both fetchers use the same arguments to order, paginatem, and filter
the package listings. Fetchers receive these arguments as a single
object, for which the interface is exported from the package. I would
have liked them to get the args separately for simplicity's and
consistency's sake, but that would have lead to lot of boilerplate code
both in thunderstore-api and in dapper-ts. Especially when more package
listing fetchers are added in the future.

Refs TS-1875