Implement composable forms logic
Break down the CreateTeamForm component to generic & composable smaller
pieces, making it easy to re-use logic across future forms.

Several TODOs are still included, notably the API error parsing/handling
needs a bit more work before it works as it's intended. That however is
the responsibility of the API library, not the components.

It's also a good idea to move the form components & logic to their own
package, but that refactoring is beyond the scope of this commit.

No refs