@thunderstore/cyberstorm: add RelativeTime
A simple component for showing a datetime in a human-friendly manner
(e.g. "4 hours ago"). Hovering the value shows the full datetime.

Use the new component on package's detail view to show original upload
time and last update time.

There's some other places where, like the PackageCard, where this could
also be used, but the MetaItem component (not to be confused with the
aptly named MetaInfoItem) accepts currently only strings, which does't
work with the "full datetime on hover" functionality. Refactoring them
is out of the scope of this commit/PR.

Refs TS-1975