@thunderstore/cyberstorm: add ImageWithFallback
This component can be used with items that have an optional image field
and thus are not guaranteed to actually have an image.

Creating a generic solution for the problem was commissioned, but it
seems currently the only use case are the communities, since packages
are guaranteed to have an associated image, and avatar use a different,
letter-based approach which is already implemented.

The images currently use 3 / 4 aspect ratio and assumes the parent sets
the width. This might not be very generic approach, but it works, at
least for now.

Refs TS-1802