@thunderstore/cyberstorm-storybook: fix MultiSelectSearch story
It appears that the typing of the component's props have changed but
the story was not updated. Consequently this broke the CI pipeline.

Hopefully the unknown type can be removed from the props once the work
on the component continues, which in turn would allow better typing for
this story.

Refs no ticket