@thunderstore/dapper: add dependency_count to PackageListingDetails
To avoid bloat on packages with a large number of dependencies, backend
initially returns up to 4 items. More can be loaded later if user so
wishes (method not implemented yet).

Total number of dependencies is returned in a separate field so UI can
indicate that the shown list of dependencies is incomplete.

Refs TS-2012