@thunderstore/thunderstore-api: update test config
The package in QA env that's used the by CI pipeline didn't have a
changelog. This wasn't initially a problem, but a code review comment
requested the backend to be changed to return 404 in such a case.

The new package has a changelog and is managed by yours truly, meaning
required changes to the test data can be more easily updated. I'd hope
to have more comprehensive test data some day, but getting distracted
by that right now seemed like a bad idea.

Refs TS-2008