@thunderstore/dapper: support deactivated dependencies
Since deactivated dependencies are still required to use a mod, backend
will return them with the rest of the dependencies from now on. New
field is_active is returned with the data. Since the reason for
deactivation might be a gross icon or description text, null and a
placeholder text is returned instead.

Refs TS-2013