@thunderstore/cyberstorm: support deactivated dependencies
To accommodate for the fact that deactivated dependencies will return
a null instead of icon URL:

- Use ImageWithFallback component instead of img element
- Add support for package icon type to ImageWithFallback
- Add support for square images to ImageWithFallback

These changes were done since it's the minimum required to prevent the
build from breaking. The following changes were considered out of scope
for this task:

- Components currently try to link to the deactivated packages, but
  their name should be shown as plain text instead
- Deactivated dependencies should have "Removed" tag on the sidebar
  of the PackageDependencyList and PackageDependencyDialog

The description for the deactivated dependencies is also censored by
the backend, but that requires no action from the client.

Refs TS-2013