Improve ApiAction and fix hook mismatch in PackageSearch re-renders
In the last tag the ability to like packages was added to PackageSearch.
This caused React issues because there were different amount of hooks
between re-renders. The solution here is to generate the hook ones and
then just change the input data to that hook.

This solution though warranted some changes to how ApiAction accepts
data. Corresponding changes are done to the PackageLikeAction.

On the same go this commit improves success and error message displaying
via toasts. The changes to useFormToaster are only done to
PackageLikeAction at this moment, but considering all the other actions
and forms are un-used right now, the same changes to other "action" and
"form" functions can be done, when they are taken into use.

Additionally UseApiAction now has return data types.