Fix a bug where apihost was removed from the session
The session implementation is still super flaky (sowwy :3), as a result when me (Oksamies) added a session clearing function to prevent non-logged in users to access the new upload page; was ran when no currentUser was detected: the apihost was removed from the session.

This commit quick fixes the issue, but more concise planning of the session architecture needs to be done.