Add PackageVersion pages and tabs (without community)
These pages are and should be exactly the same, not including community related features and functionalities.

The reason these have to be their own files is that, react-router doesn't support using the same files for two routes at the same time.

The "Required" is mostly commented out, because the endpoint isn't yet available. (even for local development testing)