Build and test all(*) packages in CI pipeline
*: building cyberstorm-storybook and testing ts-uploader is skipped as
   both processes currently fail

This will probably build some packages multiple times: as a standalone
package and as a dependency of other package that's being built. This
is acceptable tradeoff as it's easier than trying to manually keep
track of dependencies to ensure everything gets built at least once.

Update node env to v24 as this is what's used locally and some of the
tests won't work on v20. Use hard coded value rather than the one that
reads the value from main package.json, because updating it breaks
other parts of CI pipeline and fixing everything at once is outside the
scope. Update setup actions too while we're at it.