Disable coverage run in CI pipelines until a decision has been made on the current issues
As the tests and coverage checks rely on the Thunderstore repositorys tags to boot up the Django backend, that has endpoints for the test runners to use; a situation can occure where the Thunderstore repository master has commits that are not available because a new tag hasn't been made.

First occurence of this issue surfaced when adding the Package Source Tab

New PRs should preferably have tests in them, but commented out until
further notice.