@thunderstore/cyberstorm: define and export SelectOption once
Both Select and SelectSearch defined SelectOption, but only one of them
exported it. Perhaps this was an attempt to avoid naming conflict, but
in practice it forces @thunderstorm/cyberstorm-remix to either import
using a direct path (breaking the separation of packages) or use the
SelectOption defined in the sister component. Neither is ideal, so the
type is now defined once in a helper file and imported/exported from
there for everyone to use.