Add useReportPackageForm custom hook
The idea here is to further separate the sub component from the main
view component. The view component had made some optimizations that
assumed everything is located in the same file. To circumvent this,
small changes were required to its markup structure.

Some corners were cut short on the hook's side when it comes to
handling the props-as-promise. E.g. it's assumed the listing always
eventually resolves to correct values, but if that doesn't happen, the
whole page is broken anyway. Also no fallback is currently provided
while the promise is resolving (null is returned instead, but this can
be changed). useEffect is used to resolve the props rather than
Suspense and Await components, as the latters caused ~1s delay for the
modal to open, since when moved to hook, they started processing only
when the user clicked the report button.