Adjust ReportPackage form styles
I'm not entirely sure how modal widths are supposed to be adjusted, or
if they're supposed to just take the space required by the content.
Currently the "Additional information (optional)" label sets the width
for the modal, which means the textarea on the form is very narrow on
desktop, making annoying to write anything longer than a couple of
words. On the other hand setting a min width to the textarea makes it
not that responsive on mobile layouts. I think this should be addressed
in some generic fashion, perhaps in the Modal component, and is
therefore outside the scope of these changes.