Bring ReportPackage Nimbus component closer to how other components are done
The existing implementation is very well done, but it's not aligned to how other similar components are done in the project. This commit introduces an implementation that is a bit simpler, with the added negative effect of having to have two usage instances of the component in the same page (packageListing route).

Performance effects are mostly unknown, but no noticable difference was discovered in local dev env testing.