Fix package like action usage in package page
As the package like action uses a toast, it needs the provided to be
present, when it's fired up. This means that it can't be a globally "in
the file" defined function, but rather has to be defined within the
route function, so that it has access to the provider when it's
initializedFix package like action usage in package page

As the package like action uses a toast, it needs the provided to be
present, when it's fired up. This means that it can't be a globally "in
the file" defined function, but rather has to be defined within the
route function, so that it has access to the provider when it's
initialized.