Remove clientLoader from teamSettings
The component is now a dumb wrapper for showing a row of tab links.
Each tab will fetch required content from API and validate team name
and user permissions while at it.

Fetchign the team information from API to render this component was
unnecessary, as only the team name was used and it's available as a
URL parameter.