Avoid repetition in team setting tab client loaders
Team name is always returned from the clientLoader to avoid needing to
call useParams just to access it.

makeTeamSettingsTabLoader assumes params.namespaceId is always defined.
On manual testing I couldn't produce a case where this was not true, as
React Router returned 404 for any requests for URLs where it was empty
(except for profiles which act as index, as for that it's a different
route altogether).