Prepare team members component for moving to use Suspense
As part of the upcoming change, clientLoader will start returning a
Promise rather than the resolved value. For Suspense to work, the
body of the component function can't depend on the resolved values,
so move such parts into subcomponents.

Also adds link to the "team settings settings tab" for users own row
on the membership table to make it clearer how they can remove
themselves from a team.