Use API to check team leaving/disbanding permissions
There's no longer need to check if currentUser is set in the session
context as the API call will return 401 that gets handled automatically
(currently by showing 500 error but better once the global error
handling PR is merged) if the user isn't authenticated.

There's no need to wait for the teamName to resolve as it's resolved
synchronously (based on the team name in the route URL), i.e. no
Promise is returned.