Prevent team disband submission until confirmation input validates
Also allow submitting the form with enter key press on the text input
as this is the expected behaviour on web forms.

I considered trying to implement the functionality using strongForm's
refiner and onRefinerError methods but that seemed like more trouble
than it's worth, especially given that the strongForm doesn't have
internal "is disabled" state currently.